divided the header items into smaller ones
